requirements.txt
pandas
yfinance
matplotlib
schedule
python-dotenv
twilio
requests
